TIME STEP SIZE ADAPTIVITY    added explicit cast to (double) compilation issue on Hop&Malt    (thanks to Peter)
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@4281 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_timada/timeadaptivity.cpp
@@ -139,7 +139,7 @@ void TimeAdaptivity::Indicate
       break;
    case norm_rms:
       locdiserrn_->Norm2(&norm);
-      norm /= sqrt((*locdiserrn_).GlobalLength());
+      norm /= sqrt((double) (*locdiserrn_).GlobalLength());
       break;
    case norm_inf:
       locdiserrn_->NormInf(&norm);
